requirements.txt
altgraph==0.17.2
cffi==1.15.0
cryptography==36.0.2
future==0.18.2
lxml==4.8.0
packaging==21.3
pefile==2021.9.3
Pillow==9.1.0
py-enigma==0.1
pyAesCrypt==6.0.0
pycparser==2.21
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.3
pyparsing==3.0.8
PySide6==6.2.4
pywin32==303
pywin32-ctypes==0.2.0
shiboken6==6.2.4
WMI==1.5.1
